Port horizon utils tests to Python 3
* Use literal octal number syntax instead of the oct() function. The
  oct() function gives a different result on Python 2 and Python 3
  (Python 3 uses the "0o" prefix, not Python 2).
* format_value() explicitly cast the result to float. It's needed on
  Python 3 to cast a Decimal to float, because round(Decimal) returns a
  Decimal.
* tox.ini: run all horizon.test.tests.utils tests on Python 3.4

Partial-Implements: blueprint porting-python3

+++ horizon/utils/functions.py
@@ -147,4 +147,5 @@ def format_value(value):
     value = decimal.Decimal(str(value))
     if int(value) == value:
         return int(value)
-    return round(value, 1)
+    # On Python 3, an explicit cast to float is required
+    return float(round(value, 1))
